requirements.txt
fastapi
uvicorn
pydantic
requests
pytest
httpx
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
anyio
psycopg2-binary
sqlalchemy